requirements.txt
scapy>=2.5.0
pyshark>=0.6
cryptography>=41.0.0
requests>=2.31.0
python-nmap>=0.7.1
colorama>=0.4.6
rich>=13.0.0
pandas>=2.0.0
matplotlib>=3.7.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 